+++ database/migrations/2024_09_24_000009_create_emergency_contacts_table.php
@@ -12,20 +12,41 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::create('emergency_contacts', static function (Blueprint $table) {
-            $table->id();
-            $table->string('name', 191);
-            $table->string('phone', 25);
-            $table->boolean('status')->default(0);
-            $table->unsignedBigInteger('user_id');
-            $table->timestamps();
+        $tableName = 'emergency_contacts';
 
-            // Index for user_id
-            $table->index('user_id');
+        if (Schema::hasTable($tableName)) {
+            Schema::table($tableName, function (Blueprint $table) use ($tableName) {
+                if (!Schema::hasColumn($tableName, 'name')) {
+                    $table->string('name', 191)->nullable();
+                }
+                if (!Schema::hasColumn($tableName, 'phone')) {
+                    $table->string('phone', 25)->nullable();
+                }
+                if (!Schema::hasColumn($tableName, 'status')) {
+                    $table->boolean('status')->default(0)->nullable();
+                }
+                if (!Schema::hasColumn($tableName, 'user_id')) {
+                    $table->unsignedBigInteger('user_id')->nullable();
+                    // Foreign key constraint
+                    $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                }
+            });
+        } else {
+            Schema::create($tableName, function (Blueprint $table) {
+                $table->id();
+                $table->string('name', 191)->nullable();
+                $table->string('phone', 25)->nullable();
+                $table->boolean('status')->default(0)->nullable();
+                $table->unsignedBigInteger('user_id')->nullable();
+                $table->timestamps();
 
-            // Foreign key constraint
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+                // Index for user_id
+                $table->index('user_id');
+
+                // Foreign key constraint
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
